+++ CHANGELOG.md
@@ -92,3 +92,6 @@
   * [FEATURE]: Move all mutatble data properties to the data {} object
     (title, mode) **REQUIRED STEPS: The MiniReel player must be updated
     to read these properties off of the data object as well**
+  * [FIX]: Adds alert dialog before leaving a published and edited minireel
+  * [FIX]: Adds alert dialog before saving changes to a published minireel
+  * [FIX]: Changes edit card button text to 'Save' and 'Done' based on status